Added base58Check functions and then created the lib base58_nostd in the core/libs folder
